[size] fix alignment and size when using percentages
As soon as the text node is mounted we calculate the dimensions
of the textNode and adjust its fontSize according to the size of
the parent. This is much more reliable and consistent than calculating
fontSize in advance.

For percentages the old calculation would take a percentage of the
initial fontSize which is not the behaviour that is expected. We are
expected to express the fontSize in terms of the avatar dimensions.